backend/requirements.txt
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
python-decouple>=3.8
requests>=2.31
google-generativeai>=0.4.1
